Add files via upload
This mini-project implements the Caesar cipher algorithm, a classic encryption technique. It allows users to encrypt and decrypt messages by shifting letters in the alphabet. The program offers a user-friendly interface, customizable shift values, and preserves letter case. Explore the fundamentals of encryption with this project and enjoy secure message transmission